fixed variable type

+++ azheganova.faina/I6/arccos.h
@@ -19,6 +19,6 @@ class findMemberOfRow
     unsigned countofnumbers;
     unsigned countmembers;
     unsigned currentmember;
-    unsigned pi;
+    double pi;
 };
 #endif
